moved crd client to client dir
